+++ HISTORY.rst
@@ -24,6 +24,8 @@ XXXX-XX-XX
   MemoryError.  (patch by sylvainduchesne)
 - 1309_: [Linux] Process.status() is unable to recognie "idle" and "parked"
   statuses (returns '?').
+- 1313_: [Linux] disk_io_counters() can report inflated IO counters due to
+  erroneously counting base disk device and its partition(s) twice.
 
 5.4.6
 =====